+++ internal/checks/base_test.go
@@ -313,6 +313,24 @@ func mustParseContent(content string) (entries []*discovery.Entry) {
 	return entries
 }
 
+func parseWithState(content string, state discovery.ChangeType) []*discovery.Entry {
+	entries := mustParseContent(content)
+	for _, e := range entries {
+		e.State = state
+	}
+	return entries
+}
+
+func parseWithStatePath(content string, state discovery.ChangeType, name, symlinkTarget string) []*discovery.Entry {
+	entries := mustParseContent(content)
+	for _, e := range entries {
+		e.State = state
+		e.Path.Name = name
+		e.Path.SymlinkTarget = symlinkTarget
+	}
+	return entries
+}
+
 type requestCondition interface {
 	isMatch(*http.Request) bool
 }
@@ -689,9 +707,6 @@ func generateSampleWithValue(labels map[string]string, val float64) *model.Sampl
 }
 
 func generateSampleStream(labels map[string]string, from, until time.Time, step time.Duration) (s *model.SampleStream) {
-	// Truncate to seconds to ensure deterministic results when using time.Now() in tests.
-	from = from.Truncate(time.Second)
-	until = until.Truncate(time.Second)
 	if from.After(until) {
 		panic(fmt.Sprintf("generateSampleStream() got from > until: %s ~ %s", from.UTC().Format(time.RFC3339), until.UTC().Format(time.RFC3339)))
 	}